move to fs_test
